Pending commits
git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@1002 5079c8e4-9419-0410-93bb-9e5ee278c886

+++ hpx/runtime/naming/server/reply.hpp
@@ -102,7 +102,7 @@ namespace hpx { namespace naming { namespace server
                          command == command_statistics_moment2);
             
             for (std::size_t i = 0; i < command_lastcommand; ++i)
-                statistics_.push_back(f(totals[i]));
+                statistics_.push_back(double(f(totals[i])));
         }
 
         reply (agas_server_command command, naming::address addr)
